Implement server side of event notification system
In the backend system of bike-festival-2024, we need to
 send event notification to users who subscribed to the
event. However, the number of users is expected to
 grow above 7000. If some events are subscribed by a large
 percentage of users, the number of event notification we
need to send is going to be large and our server might not
 be able to perform it efficiently.

As a result, we migrate asynq into our project, which is a
 database-backed-distributed-task-queue. It provides some
ability to deal with large amount of works and assign them
 to distinct workers to perform the work at a specific
 time.

We only implement server side of the notification system in
 this project, and use Redis as our message broker, it should
 be provided in the main project of the backend system. The
 workers will be serving to send the notification to users
at a specific time defined by event_start_time.